backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic-settings==2.5.0

# QR Code
qrcode[pil]==7.4.2

# HTTP Client (for Supabase sync)
httpx==0.27.0

# PDF Generation
reportlab==4.2.0
